Remove GestureContextProvider (unused)
