--- Sources/MapboxMaps/Gestures/CameraAnimationsManagerProtocol.swift
@@ -1,13 +1,0 @@
-import Foundation
-
-internal protocol CameraAnimationsManagerProtocol: AnyObject {
-
-    func ease(to camera: CameraOptions,
-              duration: TimeInterval,
-              curve: UIView.AnimationCurve,
-              completion: AnimationCompletion?) -> Cancelable?
-
-    func cancelAnimations()
-}
-
-extension CameraAnimationsManager: CameraAnimationsManagerProtocol { }
