+++ Sources/MapboxMaps/Gestures/GestureHandlers/RotateGestureHandler.swift
@@ -5,13 +5,11 @@ import UIKit
 internal class RotateGestureHandler: GestureHandler {
 
     internal var initialAngle: CGFloat = 0.0
-    internal weak var contextProvider: GestureContextProvider!
     internal let decelerationRate = UIScrollView.DecelerationRate.normal.rawValue
     private let mapboxMap: MapboxMapProtocol
     private let cameraAnimationsManager: CameraAnimationsManagerProtocol
 
     internal init(for view: UIView,
-                  andContextProvider contextProvider: GestureContextProvider,
                   mapboxMap: MapboxMapProtocol,
                   cameraAnimationsManager: CameraAnimationsManagerProtocol) {
         self.mapboxMap = mapboxMap
@@ -21,7 +19,6 @@ internal class RotateGestureHandler: GestureHandler {
         let rotate = UIRotationGestureRecognizer(target: self, action: #selector(handleRotate(_:)))
         view.addGestureRecognizer(rotate)
         gestureRecognizer = rotate
-        self.contextProvider = contextProvider
     }
 
     @objc internal func handleRotate(_ rotate: UIRotationGestureRecognizer) {
